+++ bodygrouper/languages/spanish.lua
@@ -1,4 +1,4 @@
-﻿NAME = "Spanish"
+NAME = "Spanish"
 LANGUAGE = {
     bodygroupChanged = "You changed %s bodygroups.",
     bodygroupChangedBy = "Your bodygroups were changed by %s.",